look 1

+++ app/Http/Controllers/PhotoController.php
@@ -3,16 +3,51 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
-
 use App\Photo;
 use App\Http\Requests;
+use Illuminate\Support\Facades\Storage;
+use Intervention\Image\ImageManagerStatic as Image;
 
 class PhotoController extends Controller
 {
     public function delete(Photo $photo)
     {
-        
-        $photo->delete();
+        if($photo->main)
+        {
+            Storage::disk('public_my')->delete('upload/adverts/thumbs/' .$photo->name. '.' .$photo->ext);
+            Storage::disk('public_my')->delete($photo->path);
+            $advert_id = $photo->advert_id;
+            $photo->delete();
+            $next_file = Photo::where('advert_id', $advert_id)->first();
+            if($next_file)
+            {
+                $next_file->main = 1;
+                $next_file->save();
+                $image = Image::make($next_file->path);
+                $image->fit(290)->save('upload/adverts/thumbs/' .$next_file->name. '.' .$next_file->ext);
+            }
+        }
+
+        return redirect()->back();
+    }
+
+    public function set_main(Photo $photo)
+    {
+        if($photo->main) return redirect()->back();
+
+        $all_photos = Photo::where('advert_id', $photo->advert_id)->where('main', 1)->get();
+        If(count($all_photos))
+        {
+            foreach($all_photos as $phot):
+                Storage::disk('public_my')->delete('upload/adverts/thumbs/' .$phot->name. '.' .$phot->ext);
+                $phot->main = 0;
+                $phot->save();
+            endforeach;
+        }
+        $photo->main = 1;
+        $photo->save();
+        $image = Image::make($photo->path);
+        $image->fit(290)->save('upload/adverts/thumbs/' .$photo->name. '.' .$photo->ext);
         return redirect()->back();
     }
 }
